Encapsulate TextStyle picker to TextStylePickerView
